+++ pysmspp/block.py
@@ -4,6 +4,7 @@ from pysmspp.smspp_tools import (
     UCBlockSolver,
     InvestmentBlockTestSolver,
     InvestmentBlockSolver,
+    TSSBSolver,
 )
 from enum import IntEnum
 
@@ -1422,6 +1423,8 @@ class SMSNetwork(Block):
         fp_solution: Path | str = None,
         smspp_solver: SMSPPSolverTool | str = "auto",
         inner_block_name: str = "Block_0",
+        logging=True,
+        tracking_period=0.1,
         **kwargs,
     ):
         """
@@ -1447,8 +1450,12 @@ class SMSNetwork(Block):
 
         inner_block_name : str (default: "Block_0")
             The name of the inner block, to decide on the automatic solver to use.
+        logging : bool (default: True)
+            Whether to enable logging during optimization.
+        tracking_period : float (default: 0.1)
+            The period (in seconds) to track optimization progress when logging is enabled.
         kwargs : dict
-            The arguments to pass to the optimization function.
+            Optional arguments to pass to the solver constructor. These can include any additional parameters required by specific solvers.
         """
 
         # Map block type to default solver (for 'auto' mode)
@@ -1456,6 +1463,7 @@ class SMSNetwork(Block):
             "UCBlock": "UCBlockSolver",
             "InvestmentBlock": "InvestmentBlockTestSolver",
             "SDDPBlock": "InvestmentBlockSolver",
+            "TwoStageStochasticBlock": "TSSBSolver",
         }
 
         # Map solver names to actual solver classes
@@ -1463,6 +1471,7 @@ class SMSNetwork(Block):
             "UCBlockSolver": UCBlockSolver,
             "InvestmentBlockTestSolver": InvestmentBlockTestSolver,
             "InvestmentBlockSolver": InvestmentBlockSolver,
+            "TSSBSolver": TSSBSolver,
         }
 
         if isinstance(smspp_solver, str) and smspp_solver == "auto":
@@ -1491,4 +1500,4 @@ class SMSNetwork(Block):
             )
 
         self.to_netcdf(fp_temp, force=True)
-        return smspp_solver.optimize(**kwargs)
+        return smspp_solver.optimize(logging=logging, tracking_period=tracking_period)