+++ pysmspp/smspp_tools.py
@@ -6,6 +6,9 @@ import numpy as np
 import os
 import time
 import psutil
+import logging
+
+logger = logging.getLogger(__name__)
 
 
 class SMSPPSolverTool:
@@ -15,43 +18,63 @@ class SMSPPSolverTool:
 
     def __init__(
         self,
-        exec_file: str = "",
-        help_option: str = "-h",
+        solver_path: Path | str,
         fp_network: Path | str = None,
         configfile: Path | str = None,
-        fp_solution: Path | str = None,
         fp_log: Path | str = None,
+        fp_solution: Path | str = None,
+        configsolution: Path | str = None,
+        help_option: str = "-h",
+        **kwargs,
     ):
         """
-        Constructor for an abstract SMSPPSolverTool.
+        Constructor for an abstract SMSPPSolverTool. Option arguments coincide with the options of the SMS++ solver tools. Additional options can be passed through kwargs. Option "-o" is automatically added to kwargs with value None if not provided, to allow logging the solution.
 
         Parameters
         ----------
-        exec_file : str
-            The name of the executable file.
-        help_option : str, optional
-            The option to display the help message, by default "-h".
+        solver_path : str
+            The name or path of the executable file.
         fp_network : Path | str, optional
             Path to the SMSpp network to solve, by default None.
+            When provided, automatically the option "-p" is added to the executable call to specify the folder of the network file.
         configfile : Path | str, optional
             Path to the configuration file, by default None.
+            This option specifies the solver configuration with option "-S" when provided.
+            The folder of the configuration file is also specified automatically with option "-c" when provided.
+        fp_log : Path | str, optional
+            When provided, the solver log is saved to the specified log file, by default None.
         fp_solution : Path | str, optional
             Path to the solution file, by default None.
-        fp_log : Path | str, optional
-            Path to the log file, by default None.
+            When provided, option "-O" is added to the executable call to specify the output solution file.
+        configsolution : Path | str, optional
+            Path to the configuration solution file, by default None.
+            When provided, option "-C" is added to the executable call to specify the configuration solution file.
+        help_option : str, optional
+            The option to display the help message, by default "-h".
+        **kwargs
+            Additional keyword arguments to pass as options to the function.
+            The keys of the kwargs should be the option name, and the value should be the option value.
+            For example, if the function has an option "-x" that takes a value, the kwargs should include {"x": value}.
         """
-        self._exec_file = exec_file
+        if isinstance(solver_path, Path):
+            self._solver_path = str(solver_path.resolve())
+        else:
+            self._solver_path = str(solver_path)
         self._help_option = help_option
 
-        self.fp_network = str(Path(fp_network).resolve())
-        self.configdir = str(Path(configfile).resolve().parent.resolve())
-        self.configfile = str(Path(configfile).resolve().name)
+        self.fp_network = (
+            None if fp_network is None else str(Path(fp_network).resolve())
+        )
+        self.configfile = (
+            None if configfile is None else str(Path(configfile).resolve())
+        )
+        self.configsolution = (
+            None if configsolution is None else str(Path(configsolution).resolve())
+        )
         self.fp_log = None if fp_log is None else str(Path(fp_log).resolve())
         self.fp_solution = (
             None if fp_solution is None else str(Path(fp_solution).resolve())
         )
-        if not self.configdir.endswith("/"):
-            self.configdir += "/"
 
         self._status = None
         self._log = None
@@ -62,21 +85,54 @@ class SMSPPSolverTool:
         self._subprocess_time = None
         self._solution_time = None
         self._computational_time = None
+        self._kwargs = kwargs
+
+        if "o" not in self._kwargs:
+            self._kwargs["o"] = None
+        if "c" in self._kwargs:
+            raise ValueError(
+                "Option 'c' is reserved for the configuration file directory."
+            )
+        if "p" in self._kwargs:
+            raise ValueError("Option 'p' is reserved for the network file directory.")
 
     def calculate_executable_call(self):
         """
-        Calculate the executable call to run the solver tool.
+        Generate the standard command-line call for SMS++ solvers and can be customized by subclasses.
 
-        This method is meant to be overridden by subclasses. The base class
-        implementation calls the function provided during initialization.
-
-        Notes
-        -----
-        Subclasses override this method to return solver-specific command strings.
+        Returns
+        -------
+        list[str]
+            The command array to execute the solver.
         """
-        raise NotImplementedError(
-            "Method calculate_executable_call must be implemented in the derived class."
-        )
+        if self.configfile is None:
+            raise ValueError("configfile must be provided (non-None).")
+        if self.fp_network is None:
+            raise ValueError("fp_network must be provided (non-None).")
+        configdir, configfile = os.path.split(self.configfile)
+        networkdir, networkfile = os.path.split(self.fp_network)
+        command = [
+            self._solver_path,
+            networkfile,
+            "-S",
+            configfile,
+        ]
+        if len(configdir) > 0:
+            command += ["-c", os.path.join(configdir, "")]
+        if len(networkdir) > 0:
+            command += ["-p", os.path.join(networkdir, "")]
+        if self.fp_solution is not None:
+            command += ["-O", self.fp_solution]
+        if self.configsolution is not None:
+            command += ["-C", self.configsolution]
+
+        for option, value in self._kwargs.items():
+            if value == "" or value is None:
+                command += [f"-{option}"]
+            else:
+                command += [f"-{option}", str(value)]
+
+        return command
 
     def __repr__(self):
         """
@@ -87,9 +143,9 @@ class SMSPPSolverTool:
         str
             A formatted string showing key solver properties.
         """
-        return f"{type(self).__name__}\n\t\n\texec_file={self._exec_file}\n\tstatus={self.status}\n\tconfigfile={self.configfile}\n\tfp_network={self.fp_network}\n\tfp_solution={self.fp_solution}"
+        return f"{type(self).__name__}\n\t\n\tsolver_name={self._solver_path}\n\tstatus={self.status}\n\tconfigfile={self.configfile}\n\tfp_network={self.fp_network}\n\tfp_solution={self.fp_solution}"
 
-    def help(self, print_message=True):
+    def help(self, print_message=True, shell=False):
         """
         Print the help message of the SMS++ solver tool.
 
@@ -99,20 +155,22 @@ class SMSPPSolverTool:
         ----------
         print_message : bool, optional
             Whether to print the message, by default True.
+        shell : bool, optional
+            Whether to execute the command through the shell. Defaults to False.
 
         Returns
         -------
         The help message.
         """
         result = subprocess.run(
-            f"{self._exec_file} {self._help_option}", capture_output=True, shell=True
+            [self._solver_path, self._help_option], capture_output=True, shell=shell
         )
         msg = result.stdout.decode("utf-8") + os.linesep + result.stderr.decode("utf-8")
         if print_message:
             print(msg)
         return msg
 
-    def optimize(self, logging=True, tracking_period=0.1, **kwargs):
+    def optimize(self, logging=True, tracking_period=0.1, shell=False):
         """
         Run the SMSPP Solver tool.
 
@@ -122,12 +180,12 @@ class SMSPPSolverTool:
             When true, logging is provided, including the executable call.
         tracking_period : float
             Delay in seconds between resource usage tracking samples.
-        **kwargs
-            Additional keyword arguments to pass to the function.
+        shell : bool
+            Whether to execute the command through the shell. Defaults to False.
         """
         from pysmspp import SMSNetwork
 
-        if not Path(Path(self.configdir).joinpath(self.configfile)).exists():
+        if not Path(self.configfile).exists():
             raise FileNotFoundError(
                 f"Configuration file {self.configfile} does not exist."
             )
@@ -135,6 +193,8 @@ class SMSPPSolverTool:
             raise FileNotFoundError(f"Network file {self.fp_network} does not exist.")
 
         command = self.calculate_executable_call()
+        if shell:
+            command = " ".join(command)
 
         start_time = time.time()
         if logging:
@@ -145,7 +205,7 @@ class SMSPPSolverTool:
             stdout=subprocess.PIPE,
             stderr=subprocess.PIPE,
             text=True,
-            shell=True,
+            shell=shell,
         )
         os.set_blocking(process.stdout.fileno(), False)  # set non-blocking read
         os.set_blocking(process.stderr.fileno(), False)  # set non-blocking read
@@ -222,7 +282,7 @@ class SMSPPSolverTool:
 
         if process.returncode != 0:
             raise ValueError(
-                f"Failed to run {self._exec_file} with error log:\n{log_error}\n\nFull log:\n{self._log}"
+                f"Failed to run {self._solver_path} with error log:\n{log_error}\n\nFull log:\n{self._log}"
             )
 
         # write output to file, if option passed
@@ -251,7 +311,7 @@ class SMSPPSolverTool:
         """
         Check if the SMS++ tool is available in the PATH.
         """
-        return shutil.which(self._exec_file) is not None
+        return shutil.which(self._solver_path) is not None
 
     def parse_solver_log(self):
         """
@@ -352,47 +412,28 @@ class UCBlockSolver(SMSPPSolverTool):
 
     def __init__(
         self,
-        fp_network: Path | str = "",
-        configfile: Path | str = "",
-        fp_solution: Path | str = None,
+        solver_path: Path | str = "ucblock_solver",
+        fp_network: Path | str = None,
+        configfile: Path | str = None,
         fp_log: Path | str = None,
+        fp_solution: Path | str = None,
+        configsolution: Path | str = None,
+        help_option: str = "-h",
         **kwargs,
     ):
         """
-        Parameters
-        ----------
-        fp_network : Path | str
-            Path to the SMSpp network to solve.
-        configfile : Path | str
-            Path to the configuration file.
-        fp_solution : Path | str, optional
-            Path to the solution file, by default None.
-        fp_log : Path | str, optional
-            Path to the log file, by default None.
+        The arguments of the constructor coincide with the options of SMSPPSolverTool; see the base class for details.
         """
         super().__init__(
-            exec_file="ucblock_solver",
+            solver_path=solver_path,
             fp_network=fp_network,
             configfile=configfile,
-            fp_solution=fp_solution,
             fp_log=fp_log,
+            fp_solution=fp_solution,
+            configsolution=configsolution,
+            help_option=help_option,
+            **kwargs,
         )
-
-    def calculate_executable_call(self):
-        """
-        Generate the command-line call for the UCBlock solver.
-
-        Returns
-        -------
-        str
-            The command string to execute the ucblock_solver.
-        """
-        exec_path = (
-            f"ucblock_solver {self.fp_network} -c {self.configdir} -S {self.configfile}"
-        )
-        if self.fp_solution is not None:
-            exec_path += f" -o -O {self.fp_solution}"
-        return exec_path
 
 
 class InvestmentBlockTestSolver(SMSPPSolverTool):
@@ -402,47 +443,30 @@ class InvestmentBlockTestSolver(SMSPPSolverTool):
 
     def __init__(
         self,
-        fp_network: Path | str = "",
-        configfile: Path | str = "",
-        fp_solution: Path | str = None,
+        solver_path: Path | str = "InvestmentBlock_test",
+        fp_network: Path | str = None,
+        configfile: Path | str = None,
         fp_log: Path | str = None,
+        fp_solution: Path | str = None,
+        configsolution: Path | str = None,
+        help_option: str = "-h",
         **kwargs,
     ):
         """
-        Constructor for the InvestmentBlockTestSolver, with executable file "InvestmentBlock_test".
-
-        Parameters
-        ----------
-        fp_network : Path | str
-            Path to the SMSpp network to solve.
-        configfile : Path | str
-            Path to the configuration file.
-        fp_solution : Path | str, optional
-            Path to the solution file, by default None.
-        fp_log : Path | str, optional
-            Path to the log file, by default None.
+        The arguments of the constructor coincide with the options of SMSPPSolverTool; see the base class for details.
         """
         super().__init__(
-            exec_file="InvestmentBlock_test",
+            solver_path=solver_path,
             fp_network=fp_network,
             configfile=configfile,
-            fp_solution=fp_solution,
             fp_log=fp_log,
+            fp_solution=fp_solution,
+            configsolution=configsolution,
+            help_option=help_option,
+            **kwargs,
         )
-
-    def calculate_executable_call(self):
-        """
-        Generate the command-line call for the InvestmentBlock test solver.
-
-        Returns
-        -------
-        str
-            The command string to execute the InvestmentBlock_test solver.
-        """
-        exec_path = f"InvestmentBlock_test {self.fp_network} -c {self.configdir} -S {self.configfile} -v -o"
-        if self.fp_solution is not None:
-            exec_path += f" -O {self.fp_solution}"
-        return exec_path
+        if "v" not in self._kwargs:
+            self._kwargs["v"] = "1"
 
     def parse_solver_log(
         self,
@@ -489,47 +513,28 @@ class InvestmentBlockSolver(SMSPPSolverTool):
 
     def __init__(
         self,
-        fp_network: Path | str = "",
-        configfile: Path | str = "",
-        fp_solution: Path | str = None,
+        solver_path: Path | str = "investment_solver",
+        fp_network: Path | str = None,
+        configfile: Path | str = None,
         fp_log: Path | str = None,
+        fp_solution: Path | str = None,
+        configsolution: Path | str = None,
+        help_option: str = "-h",
         **kwargs,
     ):
         """
-        Constructor for the InvestmentBlockSolver, with executable file "investment_solver".
-
-        Parameters
-        ----------
-        fp_network : Path | str
-            Path to the SMSpp network to solve.
-        configfile : Path | str
-            Path to the configuration file.
-        fp_solution : Path | str, optional
-            Path to the solution file, by default None.
-        fp_log : Path | str, optional
-            Path to the log file, by default None.
+        The arguments of the constructor coincide with the options of SMSPPSolverTool; see the base class for details.
         """
         super().__init__(
-            exec_file="investment_solver",
+            solver_path=solver_path,
             fp_network=fp_network,
             configfile=configfile,
-            fp_solution=fp_solution,
             fp_log=fp_log,
+            fp_solution=fp_solution,
+            configsolution=configsolution,
+            help_option=help_option,
+            **kwargs,
         )
-
-    def calculate_executable_call(self):
-        """
-        Generate the command-line call for the Investment solver.
-
-        Returns
-        -------
-        str
-            The command string to execute the investment_solver.
-        """
-        exec_path = f"investment_solver {self.fp_network} -c {self.configdir} -S {self.configfile}"
-        if self.fp_solution is not None:
-            exec_path += f" -o -O {self.fp_solution}"
-        return exec_path
 
 
 class SDDPSolver(SMSPPSolverTool):
@@ -539,49 +544,28 @@ class SDDPSolver(SMSPPSolverTool):
 
     def __init__(
         self,
-        fp_network: Path | str = "",
-        configfile: Path | str = "",
-        fp_solution: Path | str = None,
+        solver_path: Path | str = "sddp_solver",
+        fp_network: Path | str = None,
+        configfile: Path | str = None,
         fp_log: Path | str = None,
+        fp_solution: Path | str = None,
+        configsolution: Path | str = None,
+        help_option: str = "-h",
         **kwargs,
     ):
         """
-        Constructor for the SDDPSolver, with executable file "sddp_solver".
-
-        Parameters
-        ----------
-        fp_network : Path | str
-            Path to the SMSpp network to solve.
-        configfile : Path | str
-            Path to the configuration file.
-        fp_solution : Path | str, optional
-            Path to the solution file, by default None.
-        fp_log : Path | str, optional
-            Path to the log file, by default None.
+        The arguments of the constructor coincide with the options of SMSPPSolverTool; see the base class for details.
         """
         super().__init__(
-            exec_file="sddp_solver",
+            solver_path=solver_path,
             fp_network=fp_network,
             configfile=configfile,
-            fp_solution=fp_solution,
             fp_log=fp_log,
+            fp_solution=fp_solution,
+            configsolution=configsolution,
+            help_option=help_option,
+            **kwargs,
         )
-
-    def calculate_executable_call(self):
-        """
-        Generate the command-line call for the SDDP solver.
-
-        Returns
-        -------
-        str
-            The command string to execute the sddp_solver.
-        """
-        exec_path = (
-            f"sddp_solver {self.fp_network} -c {self.configdir} -S {self.configfile}"
-        )
-        if self.fp_solution is not None:
-            exec_path += f" -o -O {self.fp_solution}"
-        return exec_path
 
     def parse_solver_log(
         self,
@@ -621,55 +605,35 @@ class SDDPSolver(SMSPPSolverTool):
         self._upper_bound = np.nan
 
 
-class TSSBlockSolver(SMSPPSolverTool):
+class TSSBSolver(SMSPPSolverTool):
     """
-    Class to interact with the TSSBlockSolver tool from SMS++, with name "tssb_solver".
+    Class to interact with the TSSBSolver tool from SMS++, with name "tssb_solver".
     """
 
     def __init__(
         self,
-        fp_network: Path | str = "",
-        configfile: Path | str = "",
-        fp_solution: Path | str = None,
+        solver_path: Path | str = "tssb_solver",
+        fp_network: Path | str = None,
+        configfile: Path | str = None,
         fp_log: Path | str = None,
+        fp_solution: Path | str = None,
+        configsolution: Path | str = None,
+        help_option: str = "-h",
         **kwargs,
     ):
         """
-        Constructor for the TSSBlockSolver, with executable file "tssb_solver".
-
-        Parameters
-        ----------
-        fp_network : Path | str
-            Path to the SMSpp network to solve.
-        configfile : Path | str
-            Path to the configuration file.
-        fp_solution : Path | str, optional
-            Path to the solution file, by default None.
-        fp_log : Path | str, optional
-            Path to the log file, by default None.
+        The arguments of the constructor coincide with the options of SMSPPSolverTool; see the base class for details.
         """
         super().__init__(
-            exec_file="tssb_solver",
+            solver_path=solver_path,
             fp_network=fp_network,
             configfile=configfile,
-            fp_solution=fp_solution,
             fp_log=fp_log,
+            fp_solution=fp_solution,
+            configsolution=configsolution,
+            help_option=help_option,
+            **kwargs,
         )
-
-    def calculate_executable_call(self):
-        """
-        Generate the command-line call for the TSSB solver.
-
-        Returns
-        -------
-        str
-            The command string to execute the tssb_solver.
-        """
-        fp_dir, fp_name = os.path.split(self.fp_network)
-        exec_path = f"tssb_solver {fp_name} -p {fp_dir}/ -c {self.configdir} -S {self.configfile}"
-        if self.fp_solution is not None:
-            exec_path += f" -o -O {self.fp_solution}"
-        return exec_path
 
 
 def is_smspp_installed(solvers: list[SMSPPSolverTool] = [UCBlockSolver()]) -> bool: